chore: 全 workspace clippy -D warnings 清零（~80 处 lint 修复）
工程门禁清理，覆盖 7 个 crate：

- szrsql-protocol: P1-12 慢查询日志非重入 Mutex 死锁修复（先读阈值释放锁，再重入写日志）；
  doc_lazy_continuation / doc_overindented_list_items / empty_line_after_doc_comments 修复；
  explicit_auto_deref（&mut *rt → &mut rt）；for_kv_map；map_clone → cloned/to_vec；
  type_complexity allow 标注；is_replication_mode → matches!
- szrsql-sqlite-bridge: manual_range_patterns（1|2|3 → 1..=3）；redundant_closure → value_to_json；
  needless_range_loop → iter_mut().enumerate()
- szrsql-tds-protocol: derivable_impls（AuthMode derive Default + #[default]）；
  needless_range_loop → iter_mut().enumerate()；type_complexity allow
- szrsql-oracle-bridge: if_same_then_else → || 合并；manual_range_contains；
  redundant_closure → format_value；manual_clamp → .clamp()；unnecessary_cast；type_complexity allow
- szrsql-mysql-protocol: derivable_impls；needless_borrows_for_generic_args；
  manual_strip → strip_prefix；collapsible_if；manual_ignore_case_cmp → eq_ignore_ascii_case；
  manual_div_ceil → div_ceil；unnecessary_cast；collapsible_match；needless_splitn；
  SharedTables/SharedTablesRef 类型别名消除 12 处 type_complexity
- szrsql-bin: or_insert_with → or_default；needless_splitn → split；doc_lazy_continuation；
  type_complexity allow
- szrsql-ai: explicit_auto_deref（&*catalog → &catalog）

docs: 进度行追加工程门禁清理记录

验收：cargo clippy -D warnings --workspace 0 errors；
      szrsql-protocol --lib 428 tests 全绿；
      4 个方言 crate --lib 608 tests 全绿；szrsql-ai --lib 480 tests 全绿

+++ crates/szrsql-oracle-bridge/src/ttc.rs
@@ -293,7 +293,7 @@ fn encode_number_from_str(s: &str) -> Vec<u8> {
     };
 
     let mut result = Vec::new();
-    result.push(exp_byte as u8);
+    result.push(exp_byte);
 
     // 尾数仅包含有效数字（剥离前导零），用于 base-100 分组
     let mantissa_digits: String = all_digits.chars().skip_while(|c| *c == '0').collect();
